Make a common tmux mod that is extracted from load
