+++ common/src/lib.rs
@@ -1,4 +1,5 @@
 extern crate dirs;
+extern crate libc;
 extern crate rand;
 extern crate serde;
 
@@ -6,3 +7,4 @@ pub mod args;
 pub mod first_run;
 pub mod project_paths;
 pub mod rand_names;
+pub mod tmux;
