+++ load/Cargo.toml
@@ -8,7 +8,7 @@ publish = false
 common    = { path = "../common" }
 dirs      = "2.0.2"
 libc      = "0.2.66"
-yaml-rust = { version = "0.4.3", default-features = false }
+yaml-rust = { version = "0.4.5", default-features = false }
 
 [dev-dependencies]
 rand      = "0.7.2"
