+++ load/src/lib.rs
@@ -1,19 +1,17 @@
 extern crate dirs;
-extern crate libc;
 extern crate yaml_rust;
 
 extern crate common;
 
 pub mod command;
 pub mod project;
-pub mod tmux;
 
 use args::Args;
 use command::Commands;
 use common::project_paths::project_paths;
-use common::{args, first_run};
+use common::tmux::config::Config;
+use common::{args, first_run, tmux};
 use project::parser;
-use tmux::config::Config;
 
 pub fn exec(args: Args) -> Result<(), String> {
     let project_paths = project_paths(&args);
